Galaxies --> stars

+++ code/MyModel.h
@@ -12,7 +12,7 @@ class MyModel:public DNest3::Model
 		RJObject<MyDistribution> objects;
 
 		// The model image
-		std::vector< std::vector<long double> > image;
+		std::vector< std::vector<double> > image;
 		void calculate_image();
 
 		// How many steps since image was computed from scratch
